// src/infrastructure/interfaces/establishment/establishment.response.ts
// Generated by https://quicktype.io

export interface EstablishmentResponse {
    id: string;
    name: string;
    description: string;
    ruc: string;
    address: string;
    latitude: string;
    longitude: string;
    google_address: string;
    opening_time: string;
    closing_time: string;
    validated: boolean;
    isActive: boolean;
    user: number;
    canton: string;
  }
  